+++ Code/ObjectMapping/RKObjectMappingProviderContextEntry.m
@@ -44,7 +44,7 @@
 - (NSUInteger)hash
 {
     int prime = 31;
-    int result = 1;
+    NSUInteger result = 1;
     result = prime *[self.userData hash] ? [self.mapping hash] : [self.userData hash];
     return result;
 }
